+++ test/services/postcard_service_test.dart
@@ -11,7 +11,6 @@ import 'package:nft_collection/services/indexer_service.dart';
 import 'package:nft_collection/services/tokens_service.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../gateway/postcard_api_mock.dart';
 import '../generate_mock/gateway/mock_postcard_api.mocks.dart';
 import '../generate_mock/service/mock_account_service.mocks.dart';
 import '../generate_mock/service/mock_chat_service.mocks.dart';
@@ -20,6 +19,7 @@ import '../generate_mock/service/mock_indexer_service.mocks.dart';
 import '../generate_mock/service/mock_metric_client_service.mocks.dart';
 import '../generate_mock/service/mock_tezos_service.mocks.dart';
 import '../generate_mock/service/mock_tokens_service.mocks.dart';
+import '../mock_data/postcard_api_mock.dart';
 
 void main() async {
   setUpAll(() {
